Check for stack.conf before testing stack state in stkr_delete
If a stack.conf is present then the script that defined the stack
completed successfully and it is valid to test the stack status
before deletion.

If no stack.conf exists then the stack script exited with failure:
the stack has no valid status and can be deleted without further
checks.